Separated preflights and support-bundles
