+++ gatsby-theme-marketing/gatsby-config.js
@@ -23,7 +23,7 @@ module.exports = {
     },
     {
       resolve: `gatsby-plugin-create-client-paths`,
-      options: { prefixes: [`/spec/*`, `/explore/*`] },
+      options: { prefixes: [`/explore/*`, `/preflight/*`, `/support-bundle/*`] },
     }
   ],
 }
