+++ gatsby-theme-marketing/gatsby-node.js
@@ -45,13 +45,18 @@ exports.onCreatePage = async ({ page, actions }) => {
   const { createPage } = actions
   // Make the front page match everything client side.
   // Normally your paths should be a bit more judicious.
-  if (page.path === `/spec`) {
-    page.matchPath = `/spec/*`
+  if (page.path === `/explore`) {
+    page.matchPath = `/explore/*`
     createPage(page)
   }
 
-  if (page.path === `/explore`) {
-    page.matchPath = `/explore/*`
+  if (page.path === `/preflight`) {
+    page.matchPath = `/preflight/*`
+    createPage(page)
+  }
+
+  if (page.path === `/support-bundle`) {
+    page.matchPath = `/support-bundle/*`
     createPage(page)
   }
 }
