+++ gatsby-theme-marketing/src/components/Troubleshootsh.js
@@ -123,7 +123,7 @@ class Troubleshootsh extends React.Component {
                   <div className="u-flexTabletReflow u-marginTop--50 justifyContent--center">
                     {specs && specs.length > 0 ? this.getExampleSpecs(specs, 3).map((item, i) => (
                       <div className="example-spec-block-wrapper flex1" key={`${i}-${item.slug}`}>
-                        <Link to={`/spec/${item.slug}`}>
+                        <Link to={`/preflight/${item.slug}`}>
                           <div className="example-spec-block">
                             <p className="u-fontSize--large u-fontWeight--medium u-color--biscay u-lineHeight--default u-marginBottom--10">{item.title}</p>
                             <p className="u-fontSize--normal u-color--dustyGray u-lineHeight--more">{item.description}</p>
@@ -184,7 +184,7 @@ class Troubleshootsh extends React.Component {
                   <div className="u-flexTabletReflow u-marginTop--50 justifyContent--center">
                     {specs && specs.length > 0 ? this.getExampleSpecs(specs, 3).map((item, i) => (
                       <div className="example-spec-block-wrapper flex1" key={`${i}-${item.slug}`}>
-                        <Link to={`/spec/${item.slug}`}>
+                        <Link to={`/support-bundle/${item.slug}`}>
                           <div className="example-spec-block">
                             <p className="u-fontSize--large u-fontWeight--medium u-color--biscay u-lineHeight--default u-marginBottom--10">{item.title}</p>
                             <p className="u-fontSize--normal u-color--dustyGray u-lineHeight--more">{item.description}</p>
