+++ gatsby-theme-marketing/src/components/shared/ExploreInfo.js
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { Link } from "gatsby";
+import ExploreCard from "../ExploreCard";
 
 
 export default class ExploreInfo extends React.Component {
@@ -11,17 +11,9 @@ export default class ExploreInfo extends React.Component {
         <p className="u-fontSize--largest u-fontWeight--bold u-color--biscay u-marginBottom--10"> {name} </p>
         <div className="u-borderTop--gray">
           <div className={`Info--wrapper flex ${isMobile ? "flex-column" : "flexWrap--wrap"} u-marginTop--30`}>
-            {specs && specs.map((spec, i) => {
-              return (
-                <Link to={`/spec/${spec.slug}`} className={`${isMobile ? "InfoMobile--item" : "Info--item"} flex alignItems--center`} key={`${spec.id}-${i}`}>
-                  <span className={`category-icon`} style={{ backgroundImage: `url("${spec.iconUri}")`}}> </span>
-                  <div className="flex-column u-marginLeft--12 Info--wrapper--info">
-                    <p className="u-fontSize--largest u-color--biscay u-fontWeight--bold u-lineHeight--more info-title">{spec.title}</p>
-                    <span className="u-fontSize--small u-color--tundora body-copy u-marginTop--8 info-desc">{spec.description}</span>
-                  </div>
-                </Link>
-              )
-            })}
+            {specs && specs.map((spec, i) => (
+              <ExploreCard name={name} spec={spec} i={i} isMobile={isMobile} />
+            ))}
           </div>
         </div>
       </div>
