+++ gatsby-theme-marketing/src/components/shared/KotsLinter.js
@@ -22,10 +22,12 @@ export default class KotsLinter extends React.Component {
     const {
       lintExpressions,
       className,
+      isMobile,
+      toggleLinter
     } = this.props;
 
     return (
-      <div className={`Linter-wrapper flex-column flex1 u-overflow--hidden u-marginTop--small ${className || ""}`}>
+      <div className={`${lintExpressions?.length ? "Linter-wrapper" : "EmptyLinter-wrapper"} flex-column flex1 u-overflow--hidden ${className || ""}`}>
         <div className="Linter--console flex-1-auto flex-column u-overflow--auto">
           <div className="flex flex1 flex-column alignContent--center">
             {lintExpressions?.length ?
@@ -33,14 +35,20 @@ export default class KotsLinter extends React.Component {
                 <div key={i} className={`ConsoleBlock ${error.type} flex alignSelf--flexStart`}>
                   <span className={`icon u-${error.type}ConsoleIcon flex-auto u-marginRight--small`}></span>
                   <div>
-                    {error.path && <span>{error.path}<br/></span>}
+                    {error.path && <span>{error.path}<br /></span>}
                     <span>{this.messageFor(error)}</span>
                   </div>
+                  {isMobile && <div className="AbsoluteToggleLinter-wrapper">
+                    <span className="icon toggleLinter-icon" onClick={toggleLinter} />
+                  </div>}
                 </div>
               ))
               :
               <div className="flex1 flex-column alignItems--center justifyContent--center">
-                <p className="u-fontSize--normal u-fontWeight--medium u-color--dustyGray">Everything looks good!</p>
+                <p className="u-fontSize--normal u-fontWeight--normal u-padding--12">No warnings or errors found in YAML</p>
+                {isMobile && <div className="AbsoluteToggleLinter-wrapper">
+                  <span className="icon toggleLinter-icon" onClick={toggleLinter} />
+                </div>}
               </div>
             }
           </div>
