+++ gatsby-theme-marketing/src/scss/components/TroubleshootLifecycle.scss
@@ -30,7 +30,7 @@
     }
 
     &:last-child::before {
-      height: 48px;
+      height: 100px;
     }
   }
 
@@ -86,24 +86,26 @@
     }
 
     .step {
-      width: 190px;
+      width: 255px;
       height: 72px;
       left: auto;
 
       &::before {
-        width: 100%;
-        height: 100%;
+        width: 210px;
         content: "";
         display: block;
         background-image: url("../../images/main_spritesheet.svg");
         background-repeat: no-repeat;
         background-size: initial;
-        background-position: -277px -327px;
+        background-position: -70px -317px;
         position: absolute;
-        top: -80px;
+        top: 35px;
         left: 1px;
         z-index: -1;
-        transform: rotate(-90deg);
+      }
+
+      &:last-child::before {
+        height: 15px;
       }
 
       .step-content {
@@ -114,24 +116,24 @@
         bottom: 5px;
 
         .step-content {
-          left: 20px;
-          top: 40px;
+          left: 55px;
+          top: 35px;
         }
       }
       &.customer-issue {
         bottom: 5px;
 
         .step-content {
-          right: 40px;
-          top: 38px
+          right: 70px;
+          top: 35px;
         }
       }
       &.escalation-needed {
         top: -5px;
 
         .step-content {
-          right: 40px;
-          top: 38px;
+          right: 70px;
+          top: 35px;
         }
       }
     }
