+++ gatsby-theme-marketing/src/scss/components/TroubleshootSpec.scss
@@ -13,12 +13,18 @@
   position: absolute;
 }
 
-.AbsoulteLintExpressions-wrapper {
+.LintExpressions-wrapper {
+  min-height: 44px;
+  padding: 10px;
+  background-color: #F8F8F8;
+  border: 1px solid #C5C5C5;
+  box-sizing: border-box;
+}
+
+.MobileLintExpressions-wrapper {
   position: absolute;
-  top: 324px;
-  right: 18px;
-  z-index: 10;
-  height: 408px;
+  z-index: 100;
+  height: 400px;
 }
 
 .ErrorBlock {
@@ -30,4 +36,38 @@
   font-weight: normal;
   font-size: 12px;
   line-height: 20px;
+}
+
+.EditorHeader-wrapper {
+  background: #F8F8F8;
+  border-top: 1px solid #C5C5C5;
+  border-left: 1px solid #C5C5C5;
+  border-right: 1px solid #C5C5C5;
+  box-sizing: border-box;
+  padding: 12px;
+}
+
+.AbsoluteToggleLinter-wrapper {
+  position: absolute;
+  top: 360px;
+  right: 10px;
+}
+
+.Tags-wrapper {
+  display: inline;
+  padding: 0px;
+
+  &:last-child {
+    span::after {
+      content: "";
+    }
+  }
+}
+
+.tag-card {   
+  font-family: "Helvetica Neue";
+}
+
+.comma::after{
+  content: ",";
 }
