+++ gatsby-theme-marketing/src/scss/components/shared/Linter.scss
@@ -71,6 +71,11 @@
   border: 1px solid #C4C8CA;
   background-color: #ffffff;
 }
+
+.EmptyLinter-wrapper { 
+  border: 1px dashed #C5C5C5;
+  background-color: #ffffff;
+}
 .Linter-wrapper .Linter-header {
   padding: 8px 12px;
   background-color: #F5F8FC;
