+++ gatsby-theme-marketing/src/scss/utilities/base.scss
@@ -284,12 +284,6 @@ body a {
   }
 }
 
-.AbsoulteCopyYaml--wrapper {
-  position: absolute;
-  top: 345px;
-  right: 25.75%;
-}
-
 .iframe-placeholder {
   position: relative;
   width: 100%;
