+++ gatsby-theme-marketing/src/scss/utilities/colors.scss
@@ -38,6 +38,10 @@
   color: #163166;
 }
 
+.u-color--waikawaGray {
+  color: #5C6F94;
+}
+
 .u-background--alabaster {
   background-color: #F8F8F8;
 }
