+++ gatsby-theme-marketing/src/scss/utilities/icons.scss
@@ -148,3 +148,29 @@
   background-position: 0px -235px;
   margin-right: 5px;
 }
+.icon.toggleLinter-icon {
+  width: 30px;
+  height: 30px;
+  background-position: -197px -373px;
+}
+.icon.forwardArrow-icon {
+  width: 9px;
+  height: 10px;
+  margin-top:2px;
+  background-position: 0px -256px;
+}
+.icon.preflightSmall-icon {
+  width: 16px;
+  height: 16px;
+  background-position: 0px -271px;
+}
+.icon.supportSmall-icon {
+  width: 16px;
+  height: 16px;
+  background-position: 0px -289px;
+}
+.icon.tagLight-icon {
+  width: 12px;
+  height: 12px;
+  background-position: 0px -309px;
+}
