+++ gatsby-theme-marketing/src/scss/utilities/utilities.scss
@@ -40,6 +40,7 @@
 .u-padding--row { padding: 14px 12px; }
 .u-padding--more { padding: 60px 0; }
 .u-padding--10 { padding: 10px; }
+.u-padding--12 { padding: 12px; }
 .u-padding--20 { padding: 20px; }
 .u-paddingRight--20 { padding-right: 20px; }
 .u-paddingRight--more { padding-right: 25px; }
